[RA] Restores additional network related functions.
Restores Queue_AI_Normal, Message_Input, Score_Presentation.
Restores SessionClass Create_Connections and Compute_Unique_ID.
Restores Send_Statistics_Packet.
Correct const status of PacketClass and FieldClass for string literals.

+++ redalert/session.h
@@ -39,6 +39,7 @@
 #include "connect.h"
 #include "version.h"
 #include "event.h"
+#include <stdint.h>
 
 //---------------------------------------------------------------------------
 // Forward declarations
@@ -528,7 +529,7 @@ public:
     //.....................................................................
     int Create_Connections(void);
     bool Am_I_Master(void);
-    unsigned long Compute_Unique_ID(void);
+    uint32_t Compute_Unique_ID(void);
 
     //.....................................................................
     // File I/O
